Safari-611.1.10/LayoutTests/http/tests/security/mixedContent/insecure-form-in-main-frame.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    testRunner.dumpFrameLoadCallbacks();
    testRunner.setCanOpenWindows();
    testRunner.setCloseRemainingWindowsWhenComplete(true);
}

window.addEventListener("message", function (e) {
  if (window.testRunner)
      testRunner.notifyDone();
}, false);

</script>
<p>This test opens a window that has a form with an non-https action.  We should
trigger a mixed content callback because the main frame in the window is HTTPS
but is running insecure content.</p>
<script>
onload = function() {
    window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-form.html");
}
</script>
</body>
</html>
