LayoutTests/platform/ios-wk1/http/tests/security/mixedContent/insecure-form-in-main-frame-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didCommitLoadForFrame
didDisplayInsecureContent
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This test opens a window that has a form with an non-https action. We should trigger a mixed content callback because the main frame in the window is HTTPS but is running insecure content.
